+++ GameOne/src/com/exercise/main/GameOneObject.java
@@ -14,11 +14,11 @@ public abstract class GameOneObject {
 	
 	protected ID id;
 	
-	protected int x, y;
+	protected float x, y;
 	
-	protected int velX, velY;
+	protected float velX, velY;
 	
-	public GameOneObject(int x, int y, ID id) {
+	public GameOneObject(float x, float y, ID id) {
 		this.x = x;
 		this.y = y;
 		this.id = id;
@@ -36,35 +36,35 @@ public abstract class GameOneObject {
 		this.id = id;
 	}
 
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 
-	public int getVelX() {
+	public float getVelX() {
 		return velX;
 	}
 
-	public void setVelX(int velX) {
+	public void setVelX(float velX) {
 		this.velX = velX;
 	}
 
-	public int getVelY() {
+	public float getVelY() {
 		return velY;
 	}
 
-	public void setVelY(int velY) {
+	public void setVelY(float velY) {
 		this.velY = velY;
 	}
